requirements.txt
django>=4.2,<5.0
psycopg2-binary
python-dotenv
djangorestframework
django-otp
python-decouple
